Move project related value objects to org.knora.webapi.slice.admin.domain.model
